backend/requirements.txt
flask
Flask-CORS
Flask-SQLAlchemy
SQLAlchemy-serializer
webargs
requests-html
python-dotenv
Pillow
opencv-python
gunicorn
